Fixed Breaking platform rotating after respawning
In the beginning, the hitbox of the falling platform would be deactivated as soon as it started falling. When I changed that so it was still there when it started falling a bug occured: When you hit the platform as it was falling, it started rotating and continued to do so when it respawned. Now it doesn't do that anymore.